+++ protocols/kad/src/addresses.rs
@@ -28,6 +28,7 @@ pub struct Addresses {
     addrs: SmallVec<[Multiaddr; 6]>,
 }
 
+#[allow(clippy::len_without_is_empty)]
 impl Addresses {
     /// Creates a new list of addresses.
     pub fn new(addr: Multiaddr) -> Addresses {